+++ spring-cloud-starter-tencent-polaris-router/src/test/java/com/tencent/cloud/polaris/router/endpoint/PolarisRouterEndpointTest.java
@@ -42,7 +42,8 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.when;
 
 /**
- * Test for {@link PolarisRouterEndpoint}
+ * Test for {@link PolarisRouterEndpoint}.
+ *
  * @author lepdou 2022-07-25
  */
 @RunWith(MockitoJUnitRunner.class)